gateway/requirements.txt
requests
WTForms
Flask-WTF
mysqlclient